+++ package.json
@@ -33,13 +33,13 @@
     "@types/node-fetch": "^2.5.7",
     "@types/ramda": "^0.27.14",
     "camel-case": "^4.1.1",
-    "debug": "^4.1.1",
-    "fp-ts-contrib": "^0.1.18",
+    "debug": "^4.2.0",
     "fp-ts": "^2.8.2",
+    "fp-ts-contrib": "^0.1.18",
     "imgix-core-js": "^2.3.1",
     "imgix-url-params": "^11.7.0",
-    "io-ts-reporters": "^1.2.2",
     "io-ts": "^2.2.10",
+    "io-ts-reporters": "^1.2.2",
     "jsuri": "^1.3.1",
     "node-fetch": "^2.6.0",
     "ramda": "^0.27.1",
@@ -51,14 +51,16 @@
     "@semantic-release/git": "^9.0.0",
     "@types/jest": "^26.0.15",
     "cross-env": "^7.0.2",
-    "jest": "^26.1.0",
-    "prettier": "^2.0.5",
-    "ts-jest": "^26.1.1",
-    "typescript": "^3.9.6",
-    "yarn-run-all": "^3.1.1",
+    "gatsby": "^2.26.0",
+    "gatsby-image": "^2.5.0",
     "graphql-anywhere": "^4.2.7",
     "graphql-tag": "^2.11.0",
-    "read-pkg": "^5.2.0"
+    "jest": "^26.1.0",
+    "prettier": "^2.0.5",
+    "read-pkg": "^5.2.0",
+    "ts-jest": "^26.1.1",
+    "typescript": "^3.9.6",
+    "yarn-run-all": "^3.1.1"
   },
   "release": {
     "branches": [
