+++ src/modules/gatsby-source-url/cache.ts
@@ -1,8 +1,8 @@
 import { pipe } from 'fp-ts/lib/pipeable';
-import * as TE from 'src/api/node_modules/fp-ts/lib/TaskEither';
-import { TaskEither } from 'src/api/node_modules/fp-ts/lib/TaskEither';
-import { GatsbyCache } from '../../../types/node_modules/gatsby';
-import { createLogger, trace } from './common/log';
+import * as TE from 'fp-ts/lib/TaskEither';
+import { TaskEither } from 'fp-ts/lib/TaskEither';
+import { GatsbyCache } from 'gatsby';
+import { createLogger, trace } from '../../common/log';
 
 const log = createLogger('cache');
