+++ src/modules/gatsby-source-url/gatsby-node.ts
@@ -13,15 +13,15 @@
  */
 
 import { Do } from 'fp-ts-contrib/lib/Do';
-import * as E from 'src/api/node_modules/fp-ts/lib/Either';
+import * as E from 'fp-ts/lib/Either';
 import { pipe } from 'fp-ts/lib/pipeable';
 import fs from 'fs';
-import { CreateResolversArgsPatched, GatsbyNode, PluginOptions } from '../../../types/node_modules/gatsby';
+import { CreateResolversArgsPatched, GatsbyNode, PluginOptions } from 'gatsby';
 import path from 'path';
 import readPkgUp from 'read-pkg-up';
-import { createLogger } from './common/log';
+import { createImgixClient } from '../../common/imgix-core-js-wrapper';
+import { createLogger } from '../../common/log';
 import { createRootImgixImageType } from './createRootImgixImageType';
-import { createImgixClient } from './common/imgix-core-js-wrapper';
 import { GatsbySourceUrlOptions, IGatsbySourceUrlOptions } from './publicTypes';
 
 const log = createLogger('gatsby-node');
