+++ test/unit/source-url/source-url.test.ts
@@ -5,9 +5,9 @@ import { pipe } from 'fp-ts/lib/function';
 import { CreateResolversArgsPatched, PluginOptions } from 'gatsby';
 import { FixedObject, FluidObject } from 'gatsby-image';
 import * as R from 'ramda';
-import { createLogger, trace } from '../../../src/modules/gatsby-source-url/src/common/log';
-import { createResolvers } from '../../../src/modules/gatsby-source-url/src/gatsby-node';
-import { IGatsbySourceUrlOptions } from '../../../src/modules/gatsby-source-url/src/publicTypes';
+import { createLogger, trace } from '../../../src/common/log';
+import { createResolvers } from '../../../src/modules/gatsby-source-url/gatsby-node';
+import { IGatsbySourceUrlOptions } from '../../../src/modules/gatsby-source-url/publicTypes';
 import { getSrcsetWidths } from '../../common/getSrcsetWidths';
 
 const log = createLogger('test:createResolvers');
